+++ RELEASE-NOTES.md
@@ -1,4 +1,7 @@
 ## RELEASE NOTES
+### Version 6.13.10-case-file-view-sort-by-document-upload-date
+**EUI-7812** Case file view sort by document upload date
+
 ### Version 6.13.10-case-file-view-display-document-upload-date-v3
 **EUI-7819** Case file view display document upload date